# ARROW-8113: [C++] Lighter weight variant<>