--- a/cpp/src/arrow/filesystem/mockfs.cc
+++ b/cpp/src/arrow/filesystem/mockfs.cc
@@ -50,7 +50,7 @@ struct File {
   std::string name;
   std::string data;
 
-  File(TimePoint mtime, const std::string& name) : mtime(mtime), name(name) {}
+  File(TimePoint mtime, std::string name) : mtime(mtime), name(std::move(name)) {}
 
   int64_t size() const { return static_cast<int64_t>(data.length()); }
 };
@@ -60,8 +60,18 @@ struct Directory {
   TimePoint mtime;
   std::map<std::string, std::unique_ptr<Entry>> entries;
 
-  Directory(const std::string& name, TimePoint mtime) : name(name), mtime(mtime) {}
-  Directory(Directory&&) = default;
+  Directory(std::string name, TimePoint mtime) : name(std::move(name)), mtime(mtime) {}
+  Directory(Directory&& other) noexcept
+      : name(std::move(other.name)),
+        mtime(other.mtime),
+        entries(std::move(other.entries)) {}
+
+  Directory& operator=(Directory&& other) noexcept {
+    name = std::move(other.name);
+    mtime = other.mtime;
+    entries = std::move(other.entries);
+    return *this;
+  }
 
   Entry* Find(const std::string& s) {
     auto it = entries.find(s);
@@ -90,11 +100,12 @@ struct Directory {
 };
 
 // A filesystem entry
-using EntryBase = util::variant<File, Directory>;
+using EntryBase = util::Variant<std::nullptr_t, File, Directory>;
 
 class Entry : public EntryBase {
  public:
   Entry(Entry&&) = default;
+  Entry& operator=(Entry&&) = default;
   explicit Entry(Directory&& v) : EntryBase(std::move(v)) {}
   explicit Entry(File&& v) : EntryBase(std::move(v)) {}
 
@@ -163,7 +174,7 @@ class MockFSOutputStream : public io::OutputStream {
  public:
   explicit MockFSOutputStream(File* file) : file_(file), closed_(false) {}
 
-  ~MockFSOutputStream() override {}
+  ~MockFSOutputStream() override = default;
 
   // Implement the OutputStream interface
   Status Close() override {
@@ -284,7 +295,7 @@ class MockFileSystem::Impl {
   }
 
   void GatherInfos(const FileSelector& select, const std::string& base_path,
-                   Directory& base_dir, int32_t nesting_depth,
+                   const Directory& base_dir, int32_t nesting_depth,
                    std::vector<FileInfo>* infos) {
     for (const auto& pair : base_dir.entries) {
       Entry* child = pair.second.get();
@@ -297,7 +308,7 @@ class MockFileSystem::Impl {
     }
   }
 
-  void DumpDirs(const std::string& prefix, Directory& dir,
+  void DumpDirs(const std::string& prefix, const Directory& dir,
                 std::vector<MockDirInfo>* out) {
     std::string path = prefix + dir.name;
     if (!path.empty()) {
@@ -312,7 +323,7 @@ class MockFileSystem::Impl {
     }
   }
 
-  void DumpFiles(const std::string& prefix, Directory& dir,
+  void DumpFiles(const std::string& prefix, const Directory& dir,
                  std::vector<MockFileInfo>* out) {
     std::string path = prefix + dir.name;
     if (!path.empty()) {
@@ -370,7 +381,7 @@ class MockFileSystem::Impl {
   }
 };
 
-MockFileSystem::~MockFileSystem() {}
+MockFileSystem::~MockFileSystem() = default;
 
 MockFileSystem::MockFileSystem(TimePoint current_time) {
   impl_ = std::unique_ptr<Impl>(new Impl(current_time));