… LICENSE.txt
… cpp/src/arrow/util/CMakeLists.txt
… cpp/src/arrow/util/variant_test.cc
… cpp/src/arrow/vendored/variant.hpp
… cpp/src/gandiva/literal_holder.h
… cpp/src/parquet/arrow/path_internal.cc
… go/arrow/LICENSE.txt
… python/pyarrow/includes/libarrow_flight.pxd