- [x] Unit tests for `util::Variant<>`
- [x] Compilation time and code size comparison to ensure this refactoring is beneficial
- [x] Benchmark comparison to ensure compute and nested parquet are not impacted
- [x] More docstrings and comments

Locally (release build, no ccache):

```
$ ninja
ninja: no work to do.
$ export OBJS=`ls src/arrow/CMakeFiles/arrow_objlib.dir/compute/**/*.o`

$ git checkout master && rm $OBJS && time ninja $OBJS
real    0m16.918s
user    2m41.702s
sys     0m2.495s

$ du -h --total $OBJS | tail -n 1
total 7.3M

$ git checkout 8113-Implement-a-lighter-weigh && rm $OBJS && time ninja $OBJS
real    0m14.363s
user    2m19.939s
sys     0m2.125s

$ du -h --total $OBJS | tail -n 1
6.6M    total
```