@emkornfield 2020-10-18
drive by comment.  just want to make sure we aren't using the same idioms for variant (i.e. they are compatible with C++ std)?