@bkietz 2020-10-19
@emkornfield I'm implementing a subset of `std::variant`'s functionality but all members which are present are compatible; if we switch to using `std::variant` at some point no code will break. We will need to keep the mutable-pointer visit overload since that's non standard, but IMHO that's acceptable overhead