@wesm 2020-10-20
@bkietz presumably this yields smaller code size, too?