@bkietz 2020-10-20
Updated description with build times and code sizes for a release build