@pitrou 2020-11-30
Rebased, I'm writing some micro-benchmarks.