@pitrou 2020-11-30
Variant benchmarks:
* before:
```
ConstructTrivialVariant         13290 ns        13287 ns        52532 items_per_second=752.595M/s
ConstructNonTrivialVariant     111230 ns       111215 ns         6202 items_per_second=89.916M/s
VisitTrivialVariant             25744 ns        25740 ns        27269 items_per_second=388.497M/s
VisitNonTrivialVariant          24389 ns        24385 ns        28649 items_per_second=410.082M/s
ConstructDatum                 102925 ns       102911 ns         7190 items_per_second=97.1714M/s
VisitDatum                       6204 ns         6204 ns       112708 items_per_second=1.61199G/s
```
* after:
```
ConstructTrivialVariant         13952 ns        13951 ns        50116 items_per_second=716.811M/s
ConstructNonTrivialVariant     108507 ns       108489 ns         6412 items_per_second=92.1753M/s
VisitTrivialVariant              5848 ns         5847 ns       119628 items_per_second=1.71014G/s
VisitNonTrivialVariant           5973 ns         5972 ns       118440 items_per_second=1.6745G/s
ConstructDatum                  98938 ns        98923 ns         7038 items_per_second=101.088M/s
VisitDatum                       5566 ns         5565 ns       125741 items_per_second=1.79701G/s
```