@pitrou 2020-11-30
Compute overhead benchmarks:
* before:
```
BM_CastDispatch                      971704 ns       971533 ns          698 items_per_second=1054k/s
BM_CastDispatchBaseline               97365 ns        97352 ns         6955 items_per_second=10.5185M/s
BM_AddDispatch                          165 ns          165 ns      4232825 items_per_second=6.07671M/s
BM_ExecuteScalarFunctionOnScalar    4666801 ns      4665955 ns          149 items_per_second=2.14318M/s
BM_ExecuteScalarKernelOnScalar       809563 ns       809419 ns          849 items_per_second=12.3545M/s
```
* after:
```
BM_CastDispatch                      914006 ns       913869 ns          759 items_per_second=1.12051M/s
BM_CastDispatchBaseline               95985 ns        95973 ns         7287 items_per_second=10.6697M/s
BM_AddDispatch                          167 ns          167 ns      4137185 items_per_second=5.99082M/s
BM_ExecuteScalarFunctionOnScalar    4345926 ns      4345337 ns          161 items_per_second=2.30132M/s
BM_ExecuteScalarKernelOnScalar       847525 ns       847411 ns          835 items_per_second=11.8007M/s
```