@pitrou 2020-11-30
AppVeyor seems to indicate that MSVC isn't picking up / implementing the copy constructor correctly :-(