@pitrou 2020-12-02
@bkietz You may want to review the changes I've done. I think only docs and comments remain to be done.