@pitrou 2020-12-03
I added some docstrings. I think this is ready now.