@kou 2020-12-09
I don't know why but It seems that this causes g++ 9 crash:

https://github.com/ursa-labs/crossbow/runs/1522097163#step:4:1237

```text
during GIMPLE pass: pre
In file included from /usr/include/c++/9/functional:59,
                 from /arrow/cpp/src/arrow/vendored/string_view.hpp:1480,
                 from /arrow/cpp/src/arrow/util/string_view.h:25,
                 from /arrow/cpp/src/arrow/buffer.h:31,
                 from /arrow/cpp/src/arrow/array/data.h:26,
                 from /arrow/cpp/src/arrow/array/array_base.h:26,
                 from /arrow/cpp/src/arrow/array/builder_binary.h:30,
                 from /arrow/cpp/src/arrow/compute/kernels/codegen_internal.h:26,
                 from /arrow/cpp/src/arrow/compute/kernels/codegen_internal.cc:18:
/usr/include/c++/9/bits/std_function.h: In static member function 'static void std::_Function_handler<void(_ArgTypes ...), _Functor>::_M_invoke(const std::_Any_data&, _ArgTypes&& ...) [with _Functor = arrow::compute::internal::MakeFlippedBinaryExec(arrow::compute::ArrayKernelExec)::<lambda(arrow::compute::KernelContext*, const arrow::compute::ExecBatch&, arrow::Datum*)>; _ArgTypes = {arrow::compute::KernelContext*, const arrow::compute::ExecBatch&, arrow::Datum*}]':
/usr/include/c++/9/bits/std_function.h:298:7: internal compiler error: Segmentation fault
  298 |       _M_invoke(const _Any_data& __functor, _ArgTypes&&... __args)
      |       ^~~~~~~~~
Please submit a full bug report,
with preprocessed source if appropriate.
See <file:///usr/share/doc/gcc-9/README.Bugs> for instructions.
make[2]: *** [src/arrow/CMakeFiles/arrow_objlib.dir/build.make:1610: src/arrow/CMakeFiles/arrow_objlib.dir/compute/kernels/codegen_internal.cc.o] Error 1
```

We can reproduce this on local:

```bash
(cd cpp/examples/minimal_build && docker-compose build static && docker-compose run --rm static)
```