@kou 2020-12-10
https://issues.apache.org/jira/browse/ARROW-10867 may be related.